requirements.txt
torch
numpy
comet_ml
torchvision
pyyaml
scikit-learn
fastapi
uvicorn
joblib
tensorboardX
tensorboard